Remove display tab / mode and accompanying logic and files
